clippy fixes: plan ctx mod
