+++ brahmand/src/graph_catalog/errors.rs
@@ -3,9 +3,9 @@ use thiserror::Error;
 #[derive(Debug, Clone, Error, PartialEq)]
 pub enum GraphSchemaError {
     #[error("No relationship schema found for `{rel_label}`.")]
-    RelationSchemaFound { rel_label: String },
+    RelationSchema { rel_label: String },
     #[error("No node schema found for `{node_label}`")]
-    NodeSchemaFound { node_label: String },
+    NodeSchema { node_label: String },
     #[error("No relationship index schema found for `{rel_label}`.")]
-    RelationIndexSchemaFound { rel_label: String },
+    RelationIndexSchema { rel_label: String },
 }
