+++ brahmand/src/query_planner/analyzer/analyzer_pass.rs
@@ -3,7 +3,7 @@ use std::sync::Arc;
 use crate::{
     graph_catalog::graph_schema::GraphSchema,
     query_planner::{
-        logical_plan::logical_plan::LogicalPlan, plan_ctx::plan_ctx::PlanCtx,
+        logical_plan::logical_plan::LogicalPlan, plan_ctx::PlanCtx,
         transformed::Transformed,
     },
 };
