+++ brahmand/src/query_planner/analyzer/filter_tagging.rs
@@ -9,7 +9,7 @@ use crate::query_planner::{
         AggregateFnCall, LogicalExpr, Operator, OperatorApplication, PropertyAccess, ScalarFnCall,
     },
     logical_plan::logical_plan::{Filter, LogicalPlan, ProjectionItem},
-    plan_ctx::plan_ctx::PlanCtx,
+    plan_ctx::PlanCtx,
     transformed::Transformed,
 };
 
@@ -468,7 +468,7 @@ mod tests {
         Column, Literal, PropertyAccess, TableAlias,
     };
     use crate::query_planner::logical_plan::logical_plan::{Filter, GraphNode, LogicalPlan, Scan};
-    use crate::query_planner::plan_ctx::plan_ctx::TableCtx;
+    use crate::query_planner::plan_ctx::TableCtx;
 
     fn create_property_access(table: &str, column: &str) -> LogicalExpr {
         LogicalExpr::PropertyAccessExp(PropertyAccess {
