+++ brahmand/src/query_planner/analyzer/projection_tagging.rs
@@ -11,7 +11,7 @@ use crate::{
             AggregateFnCall, Column, LogicalExpr, PropertyAccess, TableAlias,
         },
         logical_plan::logical_plan::{LogicalPlan, Projection, ProjectionItem},
-        plan_ctx::plan_ctx::PlanCtx,
+        plan_ctx::PlanCtx,
         transformed::Transformed,
     },
 };
