+++ brahmand/src/query_planner/analyzer/schema_inference.rs
@@ -9,7 +9,7 @@ use crate::{
         },
         logical_expr::logical_expr::LogicalExpr,
         logical_plan::logical_plan::{LogicalPlan, ProjectionItem, Scan},
-        plan_ctx::plan_ctx::{PlanCtx, TableCtx},
+        plan_ctx::{PlanCtx, TableCtx},
         transformed::Transformed,
     },
 };
