+++ brahmand/src/query_planner/logical_plan/plan_builder.rs
@@ -7,7 +7,7 @@ use crate::{
             errors::LogicalPlanError, logical_plan::LogicalPlan, match_clause, order_by_clause,
             return_clause, skip_n_limit_clause, where_clause,
         },
-        plan_ctx::plan_ctx::PlanCtx,
+        plan_ctx::PlanCtx,
     },
 };
