+++ brahmand/src/query_planner/optimizer/optimizer_pass.rs
@@ -2,7 +2,7 @@ use std::sync::Arc;
 
 use crate::query_planner::{
     logical_plan::logical_plan::LogicalPlan, optimizer::errors::OptimizerError,
-    plan_ctx::plan_ctx::PlanCtx, transformed::Transformed,
+    plan_ctx::PlanCtx, transformed::Transformed,
 };
 
 pub type OptimizerResult<T> = Result<T, OptimizerError>;
