+++ brahmand/src/query_planner/optimizer/projection_push_down.rs
@@ -6,7 +6,7 @@ use crate::query_planner::{
         errors::{OptimizerError, Pass},
         optimizer_pass::{OptimizerPass, OptimizerResult},
     },
-    plan_ctx::plan_ctx::PlanCtx,
+    plan_ctx::PlanCtx,
     transformed::Transformed,
 };
